% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_incomplete.R
\name{find_incomplete}
\alias{find_incomplete}
\title{Identify entities with incomplete observations (missing values)}
\usage{
find_incomplete(data, group = NULL)
}
\arguments{
\item{data}{The data frame, pdata.frame, or fixest panel object}

\item{group}{Entities' identifier (column name as character string). If NULL and data is a panel data object, the entity ID will be extracted automatically.}
}
\value{
A vector containing entities with missing values
}
\description{
Identify entities with incomplete observations (missing values)
}
\examples{
# Using base R dataset
data("airquality")
# Add a group variable
airquality$City <- rep(c("NY", "LA"), length.out = nrow(airquality))
find_incomplete(airquality, "City")

# This will now throw an error:
# find_incomplete(airquality, City)

}
